requirements.txt
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.10
django-debug-toolbar==1.11
djangorestframework==3.9.2
idna==2.8
Markdown==3.0.1
oauthlib==3.1.0
Pillow==6.2.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.8
drf-yasg
django-widget-tweaks
psycopg2
uwsgi